pyproject.toml
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "document-text-extract"
version = "0.1.0"
description = "Package to extract text from a document: PDF, docx, etc."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pymupdf>=1.26.3",
    "python-docx>=1.2.0",
]
